[GR-56861] Split CommittedMemoryProvider.
(cherry picked from commit 9ce31e3d9d91473bb19a7124139646f8c159991c)
