+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/os/CommittedMemoryProvider.java
@@ -80,39 +80,8 @@ public interface CommittedMemoryProvider {
         return VirtualMemoryProvider.get().getGranularity();
     }
 
-    @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
-    Pointer allocateAlignedChunk(UnsignedWord nbytes, UnsignedWord alignment);
-
-    Pointer allocateUnalignedChunk(UnsignedWord nbytes);
-
     Pointer allocateExecutableMemory(UnsignedWord nbytes, UnsignedWord alignment);
-
-    /**
-     * This method returns {@code true} if the memory returned by {@link #allocateUnalignedChunk} is
-     * guaranteed to be zeroed.
-     */
-    boolean areUnalignedChunksZeroed();
-
-    @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
-    void freeAlignedChunk(PointerBase start, UnsignedWord nbytes, UnsignedWord alignment);
-
-    @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
-    void freeUnalignedChunk(PointerBase start, UnsignedWord nbytes);
 
     @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
     void freeExecutableMemory(PointerBase start, UnsignedWord nbytes, UnsignedWord alignment);
-
-    /**
-     * Called by the garbage collector before a collection is started, as an opportunity to perform
-     * lazy operations, sanity checks or clean-ups.
-     */
-    default void beforeGarbageCollection() {
-    }
-
-    /**
-     * Called by the garbage collector after a collection has ended, as an opportunity to perform
-     * lazy operations, sanity checks or clean-ups.
-     */
-    default void afterGarbageCollection() {
-    }
 }
